♻️ refactor: 코드 저장 방식 개선 및 카테고리 업데이트 로직 수정, Category 컴포넌트 의존성 줄임.
